# ApiAulearning/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aulearning API",
        "description": "API de Aulearning",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/files": {
            "get": {
                "tags": [
                    "Files"
                ],
                "summary": "Listar archivos",
                "operationId": "483dc41aae90d4e129f56bbdbb646ae5",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "task_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "mime_type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "disk",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_size",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "max_size",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/PerPage"
                    },
                    {
                        "$ref": "#/components/parameters/SortBy"
                    },
                    {
                        "$ref": "#/components/parameters/SortDirection"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de archivos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Files"
                ],
                "summary": "Subir archivo",
                "description": "Sube un archivo y lo guarda en storage/app/public/tasks. Devuelve la ruta relativa y la URL pública.",
                "operationId": "b188b4e883de24b3ffb1ba36098c5a72",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "task_id"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Archivo a subir",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "task_id": {
                                        "description": "ID de la tarea asociada",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Archivo subido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Archivo subido correctamente."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "tema1.pdf"
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "tasks/abc123.pdf"
                                                },
                                                "disk": {
                                                    "type": "string",
                                                    "example": "public"
                                                },
                                                "mime_type": {
                                                    "type": "string",
                                                    "example": "application/pdf"
                                                },
                                                "size": {
                                                    "type": "integer",
                                                    "example": 152634
                                                },
                                                "task_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "url": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/storage/tasks/abc123.pdf"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/files/{id}": {
            "get": {
                "tags": [
                    "Files"
                ],
                "summary": "Ver archivo",
                "operationId": "96453911ed535b22cfebc831f23d74d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del archivo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Archivo encontrado"
                    },
                    "404": {
                        "description": "Archivo no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Files"
                ],
                "summary": "Eliminar archivo",
                "description": "Elimina el archivo físico del storage y después elimina su registro en base de datos.",
                "operationId": "af774334bc652b31c15a83bdffb37809",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del archivo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Archivo eliminado"
                    },
                    "404": {
                        "description": "Archivo no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "parameters": {
            "PerPage": {
                "name": "per_page",
                "in": "query",
                "description": "Número de elementos por página",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 15,
                    "minimum": 1
                }
            },
            "SortBy": {
                "name": "sort_by",
                "in": "query",
                "description": "Campo por el que ordenar",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": "id"
                }
            },
            "SortDirection": {
                "name": "sort_direction",
                "in": "query",
                "description": "Dirección de ordenación",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": "desc",
                    "enum": [
                        "asc",
                        "desc"
                    ]
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}